@@ -1,27 +1,33 @@
 module github.com/OpenSlides/openslides-projector-service
 
-go 1.23.1
+go 1.25.0
 
 require (
+	github.com/OpenSlides/openslides-go v0.0.0-20260209162207-733501509a02
 	github.com/caarlos0/env/v6 v6.10.1
-	github.com/goccy/go-yaml v1.14.3
-	github.com/iancoleman/strcase v0.3.0
-	github.com/jackc/pgx/v5 v5.7.1
-	github.com/redis/go-redis/v9 v9.7.0
-	github.com/rs/zerolog v1.33.0
-	golang.org/x/text v0.20.0
+	github.com/leonelquinteros/gotext v1.7.2
+	github.com/rs/zerolog v1.34.0
+	github.com/shopspring/decimal v1.4.0
+	golang.org/x/text v0.33.0
 )
 
 require (
-	github.com/cespare/xxhash/v2 v2.3.0 // indirect
-	github.com/dgryski/go-rendezvous v0.0.0-20200823014737-9f7001d12a5f // indirect
+	github.com/golang-jwt/jwt/v4 v4.5.2 // indirect
+	github.com/gomodule/redigo v1.9.3 // indirect
+	github.com/ostcar/topic v0.6.0 // indirect
+)
+
+require (
 	github.com/jackc/pgpassfile v1.0.0 // indirect
 	github.com/jackc/pgservicefile v0.0.0-20240606120523-5a60cdf6a761 // indirect
+	github.com/jackc/pgx/v5 v5.7.6 // indirect
 	github.com/jackc/puddle/v2 v2.2.2 // indirect
 	github.com/mattn/go-colorable v0.1.13 // indirect
 	github.com/mattn/go-isatty v0.0.20 // indirect
 	github.com/stretchr/testify v1.9.0 // indirect
-	golang.org/x/crypto v0.29.0 // indirect
-	golang.org/x/sync v0.9.0 // indirect
-	golang.org/x/sys v0.27.0 // indirect
+	golang.org/x/crypto v0.45.0 // indirect
+	golang.org/x/mod v0.31.0 // indirect
+	golang.org/x/sync v0.19.0 // indirect
+	golang.org/x/sys v0.39.0 // indirect
+	golang.org/x/tools v0.40.0 // indirect
 )
