@@ -1,13 +1,12 @@
 package http
 
 import (
-	"fmt"
 	"net/http"
 )
 
-func (s *ProjectorHttp) HealthHandler() http.HandlerFunc {
+func (s *projectorHttp) HealthHandler() http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		w.Header().Set("Content-Type", "application/json")
-		fmt.Fprintln(w, `{"healthy": true}`)
+		writeResponse(w, `{"healthy": true, "service":"projector"}`)
 	}
 }
