@@ -1,35 +1,50 @@
 package http
 
 import (
-	"fmt"
+	"bytes"
+	"html/template"
 	"net/http"
 	"strconv"
 )
 
-func (s *ProjectorHttp) ProjectorGetHandler() http.HandlerFunc {
+func (s *projectorHttp) ProjectorGetHandler() http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
-		// TODO: Check if user can access projector
-
 		id, err := strconv.Atoi(r.PathValue("id"))
 		if err != nil {
 			w.WriteHeader(http.StatusBadRequest)
-			fmt.Fprintln(w, `{"error": true, "msg": "Projector id invalid"}`)
+			writeResponse(w, `{"error": true, "msg": "Projector id invalid"}`)
 			return
 		}
 
-		content, err := s.Projector.GetProjectorContent(id)
+		projectorContent, err := s.projector.GetProjectorContent(id, getRequestLanguage(r))
 		if err != nil {
 			w.WriteHeader(http.StatusInternalServerError)
-			fmt.Fprintln(w, `{"error": true, "msg": "Error reading projector content"}`)
+			writeResponse(w, `{"error": true, "msg": "Error reading projector content"}`)
 			return
 		}
 
-		if content == nil {
+		if projectorContent == nil {
 			w.WriteHeader(http.StatusNotFound)
-			fmt.Fprintln(w, `{"error": true, "msg": "Projector not found"}`)
+			writeResponse(w, `{"error": true, "msg": "Projector not found"}`)
 			return
 		}
 
-		fmt.Fprintln(w, *content)
+		tmpl, err := template.ParseFiles("templates/projector.html")
+		if err != nil {
+			w.WriteHeader(http.StatusInternalServerError)
+			writeResponse(w, `{"error": true, "msg": "Error providing projector content"}`)
+			return
+		}
+
+		var content bytes.Buffer
+		if err := tmpl.Execute(&content, map[string]any{
+			"ProjectorContent": template.HTML(*projectorContent),
+		}); err != nil {
+			w.WriteHeader(http.StatusInternalServerError)
+			writeResponse(w, `{"error": true, "msg": "Error providing projector content"}`)
+			return
+		}
+
+		writeResponse(w, content.String())
 	}
 }
